transfer flatlist item to  flashcardSet card component
