+++ frontend/app/index.tsx
@@ -2,9 +2,9 @@ import { Redirect } from 'expo-router';
 import { useAuth } from '../context/AuthContext';
 
 export default function Index() {
-  const { user } = useAuth();
+  const { isAuthenticated } = useAuth();
 
-  if (!user) {
+  if (!isAuthenticated) {
     return <Redirect href="/(auth)/login" />;
   }
